+++ tests/runTests.ts
@@ -7,7 +7,9 @@ import "./fontRefreshSource.test";
 import "./migrationScripts.test";
 import "./editorUiSource.test";
 import "./homepageLayoutSource.test";
+import "./accessibilitySource.test";
 import "./localStorageUtils.test";
 import "./exportUtils.test";
 
 console.log("All utility tests passed.");
+